add contact detail

+++ src/service/account_contact_server.ts
@@ -10,9 +10,12 @@ export const getAllApps = async (options?: any) => {
     try {
         // console.log('options', options);
 
-        let query = supabase.from(db).select('*,account(id,name,avatar)');
+        let query = supabase.from(db).select('*,account:account_id(id,name,avatar)');
         if (options && options.community_id) {
             query = query.eq('community', options.community_id);
+        }
+        if (options && options.user_id) {
+            query = query.eq('owner', options.user_id);
         }
         if (options && options.category) {
             query = query.eq('category', options.category);
@@ -25,7 +28,7 @@ export const getAllApps = async (options?: any) => {
         }
 
         // const { data, error } = await query;
-        query = query.order(options?.order || 'created_at', {
+        query = query.order('account_id', { ascending: false, nullsFirst: false }).order(options?.order || 'created_at', {
             ascending: options.direction == 'asc' ? true : false
         });
 
@@ -62,13 +65,33 @@ export const createApp = async (appData: any) => {
     }
 };
 
+export const importApp = async (appData: any) => {
+    try {
+        const tasks = [supabase.from(db).insert(appData).select()];
+
+        const [createResult] = await Promise.all(tasks);
+        // console.log('collectResult', collectResult);
+
+        return {
+            success: true,
+            data: {
+                ...createResult.data
+            },
+            error: null
+        };
+    } catch (error) {
+        console.error('创建应用失败:', error);
+        return { success: false, error };
+    }
+};
+
 export const deleteApp = async (owner: string, account: string) => {
     try {
         const { data, error } = await supabase
             .from(db)
             .delete()
             .eq('owner', owner)
-            .eq('account', account);
+            .eq('account_id', account);
 
         if (error) {
             throw error;
